chore: move the object storage layer into @virtool/storage
Complete the storage half of VIR-2838. The S3 and Azure backends,
MemoryStorage, the cleanup helper, the factory, and the test fixtures
now live in the package alongside the keys, types, and errors that moved
with the scaffold.

The package owns StorageConfig. Resolving it from the environment stays
the host application's job, so apps/web's zod schema parses VT_STORAGE_*
into a value assignable to the type and the package never reads
process.env. That is what lets a second consumer supply the same shape
from wherever it keeps configuration.

Deleting the module-scope singleton is the point of the exercise: the
package now constructs nothing at import time, so it can declare
sideEffects: false. apps/web/src/server/composition.ts takes over as the
composition root and is where db and client join it in a later commit.

Tests move with the code. The package runs two Vitest projects, unit and
integration, the latter starting Garage and Azurite in testcontainers.
It gets its own CI job rather than joining Test / Packages, so pulling
those images stays out of the fast package loop.

Two upload tests stubbed uploadFileKey with a hand-written copy of the
real key format. They now mock only the singleton and let the real key
builder run — a stub that reimplements a Python-shared key is the exact
divergence keys.ts exists to prevent.
