+++ apps/web/package.json
@@ -12,10 +12,6 @@
 		"typecheck": "tsc -p tsconfig.server.json && tsc -p tsconfig.app.json"
 	},
 	"dependencies": {
-		"@aws-sdk/client-s3": "^3.1085.0",
-		"@aws-sdk/lib-storage": "^3.1094.0",
-		"@azure/identity": "^4.13.1",
-		"@azure/storage-blob": "^12.33.0",
 		"@dnd-kit/core": "^6.3.1",
 		"@dnd-kit/sortable": "^10.0.0",
 		"@dnd-kit/utilities": "^3.2.2",
@@ -81,7 +77,6 @@
 		"nitro": "3.0.260429-beta",
 		"playwright": "^1.61.1",
 		"tailwindcss": "^4.3.3",
-		"testcontainers": "^12.0.4",
 		"vite": "^8.1.5",
 		"vitest": "^4.1.10"
 	},
