+++ apps/web/src/server/analyses/data.test.ts
@@ -1,3 +1,4 @@
+import { MemoryStorage, type StorageBackend } from "@virtool/storage";
 import { eq } from "drizzle-orm";
 import { afterAll, beforeAll, beforeEach, describe, expect, it } from "vitest";
 import { seedUser } from "../auth/test/fixtures";
@@ -19,7 +20,6 @@ import { users } from "../db/schema/users";
 import { createTestDatabase, type TestDatabase } from "../db/test/fixtures";
 import { addToGroup, seedGroup } from "../groups/test/fixtures";
 import { resolveSampleActor, type SampleActor } from "../samples/data";
-import { MemoryStorage, type StorageBackend } from "../storage";
 import {
 	AnalysisIntegrityError,
 	AnalysisNoReadyIndexError,
