+++ apps/web/src/server/samples/download.test.ts
@@ -1,3 +1,4 @@
+import { MemoryStorage } from "@virtool/storage";
 import {
 	afterAll,
 	beforeAll,
@@ -7,7 +8,6 @@ import {
 	it,
 	vi,
 } from "vitest";
-
 import type { Db } from "../db/pg";
 import { takeFirstOrThrow } from "../db/rows";
 import { analyses } from "../db/schema/analyses";
@@ -24,7 +24,6 @@ import {
 import { sessions } from "../db/schema/sessions";
 import { users } from "../db/schema/users";
 import { createTestDatabase, type TestDatabase } from "../db/test/fixtures";
-import { MemoryStorage } from "../storage/memory";
 
 vi.mock("@tanstack/react-start/server", () => ({
 	deleteCookie: vi.fn(),
@@ -48,10 +47,7 @@ vi.mock("../db/pg", () => ({
 }));
 
 const storage = new MemoryStorage();
-vi.mock("../storage", async (importOriginal) => ({
-	...(await importOriginal<typeof import("../storage")>()),
-	storage,
-}));
+vi.mock("../composition", () => ({ storage }));
 
 const { handleSampleReads } = await import("./download");
 const { basicAuthHeader, seedApiKey, seedSession, seedUser, sessionCookie } =
