+++ apps/web/src/server/subtraction/download.test.ts
@@ -1,3 +1,4 @@
+import { MemoryStorage } from "@virtool/storage";
 import { eq } from "drizzle-orm";
 import {
 	afterAll,
@@ -8,7 +9,6 @@ import {
 	it,
 	vi,
 } from "vitest";
-
 import type { Db } from "../db/pg";
 import { takeFirstOrThrow } from "../db/rows";
 import { apiKeys } from "../db/schema/apiKeys";
@@ -16,7 +16,6 @@ import { sessions } from "../db/schema/sessions";
 import { subtractionFiles, subtractions } from "../db/schema/subtractions";
 import { users } from "../db/schema/users";
 import { createTestDatabase, type TestDatabase } from "../db/test/fixtures";
-import { MemoryStorage } from "../storage/memory";
 
 vi.mock("@tanstack/react-start/server", () => ({
 	deleteCookie: vi.fn(),
@@ -40,10 +39,7 @@ vi.mock("../db/pg", () => ({
 }));
 
 const storage = new MemoryStorage();
-vi.mock("../storage", async (importOriginal) => ({
-	...(await importOriginal<typeof import("../storage")>()),
-	storage,
-}));
+vi.mock("../composition", () => ({ storage }));
 
 const { handleSubtractionFile } = await import("./download");
 const { basicAuthHeader, seedApiKey, seedSession, seedUser, sessionCookie } =
