+++ apps/web/src/server/uploads/download.test.ts
@@ -1,3 +1,4 @@
+import { MemoryStorage } from "@virtool/storage";
 import {
 	afterAll,
 	beforeAll,
@@ -7,7 +8,6 @@ import {
 	it,
 	vi,
 } from "vitest";
-
 import type { Db } from "../db/pg";
 import { takeFirstOrThrow } from "../db/rows";
 import { apiKeys } from "../db/schema/apiKeys";
@@ -15,7 +15,6 @@ import { sessions } from "../db/schema/sessions";
 import { uploads } from "../db/schema/uploads";
 import { users } from "../db/schema/users";
 import { createTestDatabase, type TestDatabase } from "../db/test/fixtures";
-import { MemoryStorage } from "../storage/memory";
 
 vi.mock("@tanstack/react-start/server", () => ({
 	deleteCookie: vi.fn(),
@@ -39,10 +38,7 @@ vi.mock("../db/pg", () => ({
 }));
 
 const storage = new MemoryStorage();
-vi.mock("../storage", async (importOriginal) => ({
-	...(await importOriginal<typeof import("../storage")>()),
-	storage,
-}));
+vi.mock("../composition", () => ({ storage }));
 
 const { handleUploadDownload } = await import("./download");
 const { basicAuthHeader, seedApiKey, seedSession, seedUser, sessionCookie } =
