+++ apps/web/src/server/uploads/upload.ts
@@ -1,9 +1,9 @@
 import { UPLOAD_TYPES, type UploadType } from "@virtool/contracts";
 import { requireAuthenticatedRequest } from "../auth/middleware";
 import { hasPermission } from "../auth/policy";
+import { storage } from "../composition";
 import { db } from "../db/pg";
 import { logger } from "../logger";
-import { storage } from "../storage";
 import { createUpload } from "./data";
 
 function jsonResponse(body: unknown, status: number): Response {
