+++ packages/storage/package.json
@@ -3,7 +3,9 @@
 	"version": "0.0.0",
 	"private": true,
 	"type": "module",
+	"sideEffects": false,
 	"exports": {
+		".": "./src/index.ts",
 		"./*": "./src/*.ts"
 	},
 	"scripts": {
@@ -11,7 +13,14 @@
 		"test:watch": "vitest",
 		"typecheck": "tsc --noEmit"
 	},
+	"dependencies": {
+		"@aws-sdk/client-s3": "^3.1085.0",
+		"@aws-sdk/lib-storage": "^3.1094.0",
+		"@azure/identity": "^4.13.1",
+		"@azure/storage-blob": "^12.33.0"
+	},
 	"devDependencies": {
+		"testcontainers": "^12.0.4",
 		"vitest": "^4.1.10"
 	}
 }
